Add test:node:pretty script for prettier TAP test output using faucet
Note: When run with npm run, this runs all the tests but errors at the end with an ELIFECYCLE errno 1 (looks like permission error). To avoid the final error, the workaround is to install faucet globally (npm i -g faucet) and run it manually with npm run test:node | faucet. For practical purposes, the npm script should suffice for now.