A new middleware-friendly interface DBCore. To replace calls to indexedDB. The benefit of this will be an alternative to hooks: * Fully asynchronic write or read hooks. * Lower-level than Dexie but using Promises and bulk API.
First plan for this will be to make it possible to create middlewares supporting feaures like foreign key constraints, etc.

In the long term, the new middleware API will be simpler and a more robust alternative to hooks.

An example will be:
db.use(myMiddleware)
A middleware can optionally implement partial parts of the DBCore API and act as a proxy between Dexie and DBCore.