Repaired integration tests of Dexie.Observable by locking transaction using trans._promise instead of LockableTableMiddleware.
The difference I see is that trans._promise uses a recursive locking strategy, which may be needed for hooks because they act in a high level that once again might call down the dbcore stack and otherwise get stuck in a deadlock.

This would not be the case for pure middlewares where the middlewares should NOT call on Dexie methods again because it would possibly cause stack overflow.