Revert to Dexie 2.0 behavior regarding native await handling to potentially resolve #928.
In PR #915 did begin to always be prepared for native await calls in the scopeFunc of db.transaction(). So that it should work also if given scopeFunc was not declared async itself, but was forwarding the call to a real async function.

This commit will revert that particular part of the #928 PR. Still we are CSP compliant and all else fixed in #928, but we revert to requiring the callback function in Dexie.transaction() and Verison.upgrade() to be declared async in order to keep Dexie.currentTransaction correct between await calls.