Allow declaration-only tables for TS users. Due to how @babel/plugin-transform-typescript transpiles class fields, we had to change the Typescript recommendation of how to subclass Dexie. This commit makes it possible to go back to the previous recommendation in the docs - do only declare the classes and not have to manually initialize the tables in the constructor.
Plan: Wait with the docs updates until this commit has been released and then for another while until enough people have upgraded their Dexie to that version.

The docs should then be updated to propose a plain declaration of table properties, with an exclamation mark to hint about that the field is there despite not being assigned to in the constructor.