Collection.subscribe()
Make Collection & Table comply to es-observable's subscribe method.

https://github.com/tc39/proposal-observable

Should be compatible with RxJS. Needs to be verified.

Idea is to enable live queries without having to go with dexie-observable, since that addon affects performance by adding the mutations to a _changes table - which we really do not need here.

The limitation of this built-in observable capability is currently that it does not react ot DB mutations on other tabs/windows. However, is this really desired??? Do we want inactive tabs to start working and take performance from the computer when things happen in DB? If really needed, that functionality could be easily patched on top of this library by firing the new db.on.mutate event when storage event happens. Or in the case of a worker mutating data, it could be patched by letting the worker post message to frontend code that would fire the db.on.mutate event.