Fix UX issue with sorting order in Vue example
Because we retrieve the todos on each change to the list, if we are in
"unfinished first" order, then if someone checks off a todo at the top
of the list, that will bounce down to the bottom of the list. This
sudden change in order would reduce accessibility.

To fix this, this commit has the retrieved todos sorted to preserve
their order unless the sort order was explicitly changed.