Fine grained index based observability: * If having used get or bulkGet() of a few items (less than 15) and then do put or bulkPut() on those directly on same transaction, observabiilty will be able to detect what changes have been made on indexes.
What we want to improve is that when calling Table.update(id, changes), we should not require all index based queries to re-run - instead only those queries that queries a range who's been mutated.

With this commit, a query like db.friends.where('age').between(10,20) will not have to rerun if we:
* Delete a friend that had age 21.
* Add a friend with age 8.
* Update a friend with age 8 or 21.
But should re-run if we:
* Delete a friend with age 15
* Add a friend with age 15
* Update a friend's age from 9 to 10.
* Update a friend's age from 20 to 30.

Of course this also apply to all other type of ranges, such as equals, startsWith, above, below etc.

We assume open ranges as a compromise.

We stop recording changes if they are more than 50 and assume indexes as a whole is considered updated.