Bugfix: Must track both indexes and result keys when observing index-based queries.
Optimization: Let caching on transaction not be a generic behavior for get() and getMany(). It could slow down use cases where simple get requests are used and not for updating them.

Instead, let the Table.update() and Collection.modify() take advantage of the transaction-caching to not having to request the same entries twice. They both do it for internal purpose and never leave the getMany() result out to end user so they can rely on immutable caching so that observation-middleware can track changes better and so that hooks do not need to request the same set of objects twice.