dbName follows dependencies.indexedDB (II) (#1194)
A lighter implementation of PR #820.

Regard each indexedDB dependency as it's own "world".
Switching Dexie.dependencies.indexedDB to another implementation should enumerate databases there and not regard databases created elsewhere to listed there.
Creating a Dexie instance with {indexedDB, IDBKeyRange} set in options should act in that "world" if database is created or deleted.
Hopefully everything from PR #820 is included in this PR. I wanted to implement it more lightway without the need of classes as that could add unnescessary size to dexie.min.js. Hopefully this implementation is also easy to understand.

Note: The case when indexedDB isn't present or is behind a security policy was previously covered by a try...catch in initializeDbEnumerator() and checked for undefined whenever accessing that global variable. In this implementation we create the database enumerator on-demand, which makes this situation impossible to occur since it will only be created if a database was successfully added or deleted.

Also:
* Don't let idbAndPromise-database disturb the test for getDatabaseNames