Resolves #1225. An index-only query would not wake up unless we trigger the index. The >=50 items case did only trigger primary key and count queries.
The condition where the bug was triggered was not as I thought on outbound open range queries only, but on those subscriptions whose previous result was empty! Why?! Because those who had a non-empty previous result did also subscribe to primaryKey ranges on the last returned set of primary keys and the "catch all" case when >=50 records did put a FULL_RANGE there. This FULL_RANGE affected any primary key subscribed to, but if previous result was empty, it did not trigger.

Had first the thought to solve this by extending the subscription of previous result with adding a dummy key there, such as Infinity (which is not a normal primary key anyway). However, I decided to solve it in the mutator instead as it would be easier to understand.