dexie-cloud-addon fix Websocket reconnect issue.
WebSocket connection dies when leaving app open for more than 1 hour.

Websocket connection SHALL die after 5 minutes, BUT as soon as the user becomes active, it shall reconnect.

Problem was that there were a filter in the reconnect process that stopped it in case access token was expired.

Now, we go down the normal flow instead , which will indeed fail in the constructor of WSConnection when it calls the connect() method, which will trigger a TokenExpiredError error to the subscriber, which will be catched in connectWebSocket.ts' catchError() which will refresh token and remap the observable to a fresh observable so the subscriber gets the next() call as if nothing had happened.