Official event Dexie.on('storagemutated') that also propagate as a CustomEvent on window as "x-storagemutated-1".
This change is a step to making dexie@3.2.0 an officially supported stable release where the observability communication channel can be official and documented.

Breaking Change of undocumented event name: The renaming of the event is a "breaking change" from previous beta versions (dexie@3.1.x-beta.x and dexie@3.2.x-beta.x) even thought the event hasn't been any part of the officially supported API surface before. The name of the BroadcastChannel and storage event used is also changed with this commit. Now we're using "x-storagemutated-1" as the BroadcastChannel event, and in case storage event is used, we use a localStorage property with that name there as well. Currently the use of "storage" event is not working on Safari as it has a recent bug. However, the bug has been resolved in webkit and a future version of Safari will still gain from this communication channel in case it still do not support BroadcastChannel.

The event was previously undocumented and had the name Dexie.on('txcommitted'). The event is the channel that communicates changes on indexeddb databases to enable `liveQuery()` to detect changes across service worker, windows and tabs. Event is propagated across peers using BroadcastChannel on browsers supporting it, and "storage" event on IE and Safari.

This commit also improves observability in a rare use case when "dexie" module is imported multiple times (possibly with different versions of the package) as it also propagates on window/self event target and not just Dexie static.