Official event Dexie.on('storagemutated') (#1384)
Official event Dexie.on('storagemutated').

This change is a step to making dexie@3.2.0 an officially supported stable release where the observability communication channel can be official and documented.

The event was previously undocumented and had the name Dexie.on('txcommitted'). The event is the channel that communicates changes on indexeddb databases to enable `liveQuery()` to detect changes across service worker, windows and tabs. Event is propagated across peers using BroadcastChannel on browsers supporting it, and "storage" event on IE and Safari.

The commit does not break any officially documented Dexie API but as the rename of the "txcommitted" event to "storagemutated" and its corresponding event names in BroadcastChannel land localStorage ("x-storagemutated-1") are changed, code that would have been using those undocumented event names will need to be updated.

This commit also improves observability in a rare use case when "dexie" module is imported multiple times (possibly with different versions of the package) as it also propagates on window/self event target and not just Dexie static.