Bulk update (#1633)
* bulkUpdate()
+ support for synking intended updates in dexie-cloud-addon instead of just syncing the actual changes being made locally.

A flaw in dbcore allowed optional property 'changeSpecs' without the option to combine them with the provided keys (and not the keys sent to bulk 'put' operation, which can be fewer than the intended keys sent to table.bulkPut() because we only put keys that existed in the db currently)

We had to rename `DBCorePutRequest.changeSpecs`. Otherwise an old dexie-cloud-addon combined with a new dexie could apply the changeSpecs on the wrong keys. Making sure the interopability will be the following:

* old dexie + new dexie-cloud-addon: Work as before ('changeSpecs' is not used in old dexie)

* new dexie + old dexie-cloud-addon: Work as before (new dexie's bulkUpdate uses new property 'updates' and don't start using 'changeSpecs'

* new dexie + new dexie-cloud addon: Handle bulkUpdate() similar to how multiple calls to table.update() would behave - the indented keys and their changeSpecs are synced so that the server can reexecute the original intention on server-side data (as described in https://dexie.org/cloud/docs/consistency)

* Added a test to verify middlewares are called even when there's nothing to put.

* Fix - integration tests failing when addons mutate db on open

* Typo fix in tests-table.js

Closes #632