Revert to dexie 3 style Table<T, TKey> but allow a 3rd optional parameter InsertType.
With inspiration from @nlaurie's PR's #1632 and #1635 and descriptions of the problems he faced with 4.0.0-alpha.4.

A complex type system can create hard-to-solve typing issues and trigger subtle bugs in the typescript compiler. It's also not good for backward compability with dexie@3.2.

TKey is now a plain key again, as in dexie 3.x.

To get the power from dexie@4.0.0-alpha.4's  Table<Entity, 'id'>:
use the new generic EntityTable<T, PrimaryKeyProp>.

Using EntityTable instead of Table will opt-in to sugaring the InsertType by removing methods and making the primary key property optional in calls to table.add(), put(), bulkAdd(), bulkPut().

This is a suggestion for a solution of the problem discussed in #1632