On login, don't just spawn new WebSocket directly. Instead, wait until we also have the first pull for the new user stored so that we can compute realmSetHash correctly and avoid getting the realms-changed event from server becuase that seems to be what triggered the extra redundant pull request sent after logins.
One worry still with this change is that the old websocket stays active
and may get the realmschanged event to it.

Idea to fix also that: change from filter() in the added code to map()
and map userLogin to null as long as we lack the new state
(and if so, dont take(1) but fix it some other way.