on('ready') subscribers execute in liveQuery PSD ...if db.open() was fired as a pseudo-operation in background when a liveQuery is executed the first time.
This resulted in that on-ready callbacks execute in the liveQuery context and get exception for doing a 'rw' transaction within the live query if the on-ready callback does some write transaction.

Solution is to always make sure that db.open() executes on the top scope no matter the scope of the caller.