Dexie cloud fixes (#1717)
* Eslint disable import first warning

* CustomEvent not perfect way to communicate to self
in a service worker because they fail to addEventListener on self if
they do it ad-hoc. THey must only call self.addEventListener() from
their main script.

This change mimics built-in event handling by implementing it inline.

* On login, don't just spawn new WebSocket directly.
Instead, wait until we also have the first pull for the new user stored
so that we can compute realmSetHash correctly
and avoid getting the realms-changed event from server
becuase that seems to be what triggered the extra redundant pull request
sent after logins.

One worry still with this change is that the old websocket stays active
and may get the realmschanged event to it.

Idea to fix also that: change from filter() in the added code to map()
and map userLogin to null as long as we lack the new state
(and if so, dont take(1) but fix it some other way.

* New event db.cloud.events.syncComplete
Fullfills another purpose than syncStatus, which is a bit more complex.
The new event simply fires whenever a sync request has been completed
successfullly.

* Type correction.
computeSyncState gets change-based data and computes the new sync state on top of that. This is an area where we would need to look more into how to improve it. We are missing the previous state when computing the new state.

* dexie-cloud-addon: bump version

* dexie-cloud-addon: bugfix
"Failed to sync client changes TypeError: Cannot read properties of undefined (reading 'slice')

* .

* dexie-cloud-addon: version bump

* Merge indexes of dexie-cloud access control tables
...if also declared in application code.

* Dexie cloud addon version bump

* Upgraded deps (minor) for dexie-cloud-todo-app

* dexie-cloud-addon: Let umd module target es2016

* Bump dexie-cloud-addon version

* Updated lockfile

* dexie-cloud-todo-app: Upgraded dexie-cloud-addon

* Send realmSetHash along with ready event

* Don't send "ready" msg before resync is complete

* Bump dexie-cloud-addon

* Upgraded dep in dexie-cloud-todo-app

* Committing pnpm-lock