Dexie Cloud: Allow DB owners impersonate users. To login as other user, the following format will be understood by the service: "yourEmail as targetEmail" Example: "david@dexie.org as foo@example.com"
Allow this format in the login dialog. Server will validate whether the first email owns an API client with GLOBAL_READ and GLOBAL_WRITE access. If so, OTP will be sent to owner email but token will be given out with the access for the second email.

This will help reproducting issues and troubleshooting. Note that this is not a secutiry concern - the first user has full control over the database and access to read and manipulate the full database via REST or import/export anyway. This is just a tool to be able to reproduce customer issues by logging in as them and follow repro steps.