Never throw VersionError (#1880)
* Allow downgrade + schema manip without version update

* It should work having two versions of the DB opened at the same time as long as they have a compatible schema.

A compatible schema does not need to have the same tables or indexes. They might differ quite a lot.
* Attaching an upgrader to one of them would mean that the upgrader runs even though the first one didn't request it. Therefore, this scenario should be avoided when having upgraders that modify content between versions.
* Schemas are not compatible if they index the same property or properties using different flags (unique or multiEntry)
* Schemas are not compatible if they have different primary keys

However, in most scenarios where we would fail opening one of the DBs, we will now succeed and be able to access the same data.

* Fixing a unit test

* Support infinite number of schema changes without version bump
(Not just 10)
When native verno is incremented 10 times to the next 10 multiple, create a table $meta that contains the virtual version. Maintain that table and delete it when no longer needed.

* Skip 2 tests in Dexie.Syncable's / Observable's integration tests

* Indentation only...