Use TInsertType in Table.update and Collection.modify (#1764)
This is a typing change that opens up for library users to solve issues with the complex mapped type `UpdateSpec<T>` expected in Table.update() and Collection.modify(). In case T is a complex class with recursive references, the UpdateSpec might fail due to exhausting the typescript compiler. Now, this can be avoided by providing a simpler version of T to TInsertType argument of `Table<T, TKey, TInsertType>`.

The original intention of TInsertType is to be the POJO representation of an object that can be passed to Table.add(), Table.put() and their corresponding bulk methods. The type could for example have the primary key optional if it is auto-incremented, while the returning type T has it required.

With this PR, the TInsertType is also reused by Table.update() and Collection.modify() through `UpdateSpec<TInsertType>` to compute all possible keypaths that could be updated without risking the compiler having to traverse an endlessly recursive path in case T is a complex object with recursive properties.

See also our discussion around this in the original PR #1764.

---------

Co-authored-by: Nick <nlaurie@metafig.com>