Allow advanced modifications of properties in Collection.modify (#1910)
Add a new class PropModification with capabilities of modifying properties into computed values based on the previous value, to improve the Operation-based CRDTs support that is already possible with where clauses and `Collection.modify()` but with more precise operations than just replacing a property to another constant value.

The first operation to support is `replacePrefix` that replaces the leading string of a string. This PR also exports a helper function `replacePrefix()` to use in modify operations that returns an instance of PropModification configured to do this operation. Future operations could include `increment` (for numbers), `push` (for arrays), `add` (for unique arrays / sets), `remove` for arrays, and `updateArrayItem` for arrays with object items.

The `replacePrefix` operation could seem a bit special-casey, but the reality is that it solves an important pattern when working with tree structures and letting an indexed property represent the tree path. Using this new operation, it becomes possible to move an entire sub tree in a single modify operation:

```js
db.files
  .where('path')
  .startsWith('old/path')
  .modify({
    path: replacePrefix('old/path', 'new/other/path')
  });
```

Even though this can already be accomplished using a JS callback though doing the equivalent operation, it would not fully propagate the operation to DBCore in order to implement CRDT capabilities:

```js
db.files
  .where('path')
  .startsWith('old/path')
  .modify(file => {
    file.path = 'new/other/path' + file.path.substring('old/path'.length);
  });
```

Using this JS function would not propagate the operation to DBCore because a JS function cannot be interpreted safely for several reasons (security but also missing closures) - so sync implementations would only get the resulting individual put() operations that resulted locally from the modify operation, and would therefore not be able to consistently sync the tree move operation with the server in a truly consistent manner that would resolve correctly across multiple clients in case other clients have added, moved or deleted nodes while being temporarily offline.

Adding this new operation will allow moving trees and propagate the operation consistently in synced environments that supports and propagate the `criteria` and `changeSpec` properties in the resulting put operation to DBCore and further down into sync implementations. Currently only Dexie Cloud Server will respect this and handle it consistently, but the information is available in DBCore to implement in other endpoints.